requirements.txt
pandas 
chromadb 
ollama
google-genai 
python-dotenv
streamlit
flask
flask-cors